backend/requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.1.18
chardet==3.0.4
cloudpickle==0.6.1
cryptography==2.1.4
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.2
djangorestframework==3.9.0
entrypoints==0.2.3
httplib2==0.9.2
idna==2.6
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
Pillow==5.1.0
pip==18.1
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.2
pyRFC3339==1.0
pytz==2018.5
PyWavelets==1.0.1
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.1
scipy==1.1.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
setuptools==39.1.0